requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
click==8.1.7
APScheduler==3.10.4
